+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 All Notable changes to `ElementFinder` will be documented in this file
 
-## Unreleased [0.1.0-alpha.2 2016-05-26]
+## Unreleased [0.1.0-alpha.3]
+
+## 0.1.0-alpha.2 [2016-05-25]
 
 ### Added
 - Added `ElementFinder::query()` as an alias of `ElementFinder::node()`
@@ -11,8 +13,8 @@ All Notable changes to `ElementFinder` will be documented in this file
  
 ### Deprecated
 - #29 `ElementFinder::getNodeItems()`
-- #28 Method `ElementFinder::elements()` has been renamed to `ElementFinder::element()`, 
-- #28 Method `ElementFinder::html()` has been renamed to `ElementFinder::content()`, 
+- #28 Method `ElementFinder::elements()` has been renamed to `ElementFinder::element()`,
+- #28 Method `ElementFinder::html()` has been renamed to `ElementFinder::content()`,
 - #28 Method `ElementFinder::query()` has been renamed to `ElementFinder::executeQuery()`,
 - #28 Method `NodeHelper::getOuterHtml()` has been renamed to `NodeHelper::getOuterContent()`,
 - #28 Method `NodeHelper::getInnerHtml()` has been renamed to `NodeHelper::getInnerContent()`,
@@ -20,6 +22,7 @@ All Notable changes to `ElementFinder` will be documented in this file
 - #14 Remove 3 parameter inside `ElementFinder::KeyValue()`
 
 ## Version 0.0.3
+
 ### Changed
 - Feature #4 Use `DOMAttr::nodeValue` instead of `DOMAttr::value`
 - BC #7 Refactor `Helper` class. Create `FormHelper`, `NodeHelper` and `StringHelper`